fix: pods has warnings error after deployment - move integration plugin into the module/core - add axios proxy error handling M2-1050
