+++ packages/theme/modules/core/integrationPlugin/_proxyUtils.ts
@@ -1,15 +1,19 @@
 import { Context as NuxtContext } from '@nuxt/types';
 import { merge } from 'lodash-es';
+import { Logger } from '~/helpers/logger';
 
 export type ApiClientMethod = (...args: any) => Promise<any>;
 
 interface CreateProxiedApiParams {
+  nuxtCtx: NuxtContext;
   givenApi: Record<string, ApiClientMethod>;
   client: any;
   tag: string;
 }
 
-export const createProxiedApi = ({ givenApi, client, tag }: CreateProxiedApiParams) => new Proxy(givenApi, {
+export const createProxiedApi = ({
+  givenApi, client, tag, nuxtCtx,
+}: CreateProxiedApiParams) => new Proxy(givenApi, {
   get: (target, prop, receiver) => {
     const functionName = String(prop);
     if (Reflect.has(target, functionName)) {
@@ -19,7 +23,11 @@ export const createProxiedApi = ({ givenApi, client, tag }: CreateProxiedApiPara
     // eslint-disable-next-line @typescript-eslint/require-await
     return async (...args) => client
       .post(`/${tag}/${functionName}`, args)
-      .then((r) => r.data);
+      .then((r) => r.data)
+      .catch((err) => {
+        Logger.debug(err);
+        return nuxtCtx.error({ statusCode: err.statusCode, message: err });
+      });
   },
 });
